extract CLI into separate entry point file
